Changed folder layout for making releases easier.
Moved the source files from the project folder, to the root folder. This will make it easier to when releasing the game (through script, or by hand) where you only need to package the Game folder.
The reason why you don't want to package the src folder with the game is that the source is usless outside collaborative work, and takes 4-5 times the size of the binary files.

Moved /Game_Data/source to /src
Moved /Game_Data/utility to /utility
Renamed /Game_Data /Game

Updated /utility/config.yaml to point to /Game/Data