Added state trackers for (1) Player states and (2) Map Battle event.
If a battler targets another event, it will now set the switch ON for the switch defined in Map::State::BATTLE
When the player state changes (e.g. knockedout) it will trigger a switch on/off as defined in StateTracker.

Added a hook module to make it easier with duckpunching. Requires that the class is defined, and HookExtension is included in the class.
Use
  include HookExtension # add the ability to hook
  # register a hook for method method_name
  register_hook :method_name do |callee, return_value, *args|
    # callee = the method's object
    # return_value = the value returned from the call
    # *args = arguments passed to the method
  end