Interpreter   call_common_events_by_name(event_name) -calls all common events with the given name   get_common_events - retrieves all common events with the given name (calls index if no index is found)   index_common_events - creates an index of all common events based on their name   get_event_by_name -> get_events_by_name - changed to signal that an array is given (rather than one event)   get_event_by_index - as an alternative to using names   call_event - now get's the page of the event and relies on call_page to pass it on for interpretation   call_page - given one or more pages will now retrieve the commands from the page and pass it on to call_command   call_command - set's one or more commands as the workload of the interpreter Game_Event   page - the current page is now publically readable   condition_met?(page)- if the condtions ofthe page are met (set in editor)   get_page_by_index(index, meet_conditions) - gets the page at the provided index, does checks if conditions are met if requested (TRUE)   get_page_by_name(page_name, meet_conditions) - gets the pages that has a comment with the format "name:page_name" e.g. "name:my_page" that has the passed name. Checks conditions if requested (TRUE)   index_page_names- indexes the name of the pages. Checks the 6 first commands of each page by default for a name. This can be modified by altering the const INDEX_DEPTH_MAX.
Events overwriting the knockout CE are now required to be named "Knockout Manager"
Changed the name Knockout -> Knockout Manager on several maps

When the rape skill is executed, it will now first run CE45.
If player is nude:
  if no event named "Knockout Manager" exists on map:
    run CE300 - Default Rape Manager
  end
end

CE300 in turn checks the name of the enemy and looks for events on the map or as a CE. E.g. for a goblin:
if no event named "Rape:Goblin":
  find and run common events named "Rape:Goblin"
end

Rape:Goblin (CE303) in turn plays out the scene, as if it were an event on a map.

Example of overwritten behaviour can be found on map 33 - Bramble Woods _junction [KTS]


Fixed issue with persona showing up when it shouldn't (due to using switches to set visible state in update persona 2)
  Hide Persona -> Switch 127 = on
  Show Persona -> Switch 127 = off