Added fix for setting direction with graphics (i.e. using the fist row).
Some strange issue with the first row of the spriteset not setting the direction correctly. Managed to find the issue in the custom movement part of Game_Character, but has been unable to figure out how it is actually impacting the game as it does.

The changes are done to command 41 (change graphics) where the check if the direction passed differs from the original one before making modifications, was removed.