Packaged rxdatav with OCRA.
Changed from using ruby2exe to ocra for packaging rxdatav. Hopefully this means that users with non latin1 usernames will have an easier time with getting it to work.

Orca didn't capture all of the scripts in one executable, making us store them in a seperate folder. This means that we may actually be able to work on the script properly.