Fixed issue with scientist not being removed.
- Added empty page with condition self-switch A.
- Replaced "erase event" with "set self-switch A ON"

Erased events will only work for the duration of the time the player is on the map. Once the player leaves and returns, the event that was removed will have returned.

This is because the state of the maps aren't save in the save games.

What is saved on the other hand are the self-switches, so a solution for this issue is to set a self-switch that informs the game not to use the event you want removed.