Added script for when user doesn't run Game.bat
If the player tries to run Game.exe as the first thing they do, they'll
now get a message and the game will try to run Game.bat. If Game.bat can't
be found, the window will closed after providing an error message.

I was unable to get the linebreaks (newline) to work for some reason in
the error messages. I'm not sure if the alert window can't handled escaped
characters, or if I'm just doing something wrong.

If I write "foo\nbar" I'm expecting "foo
bar" not "foo\nbar" that I'm not getting.