Added dialog for Alice and Billy's bookcase.
The player will now have some text if the visit the bookshelf in the second floor of the farmhouse. The text depends on the context of the library book quest, guiding the player to visit Nicky (if they haven't) at the start of the quest, and direct them to Billy if they've spoken to Nicky. If they got the book, it will use the default text.

There may be a conflict if the player has had sex with Billy before, where the default dialog may show up when interacting with the bookshelf. Hopefully it's generic enough, and highlight that the book can't be found at this location.