Added the ability to set Game Over image on-the-fly.
Setting the Game Over image can now be done by calling Scene_Gameover.set_picture(filename).
Filename is a string of the name of a file found in the Graphics\Gameovers folder. Not sure why there's a folder dedicated to maximum one image (as it seems to be supported by the game).

This static method is short for modifying the $data_system object by hand.