Added fix for player routing in action triggered events.
Added a check to see if the player is facing an enemy before triggering the attack action. This is to fix the issue with events with routing being broken due to the attack animation. Instead of playing the animation the player will use the default sprite when interacting with normal events on the map.