Changed transport location from Goblin Cave to World Map.
Move the exit location from the Goblin Cave so the player doesn't get stuck on the new world map. The world map was also modified to have have an entrance back to the Goblin Cave, like the old map used to have. The shrine was moved further to the east, so it wouldn't be right on top of the new entrance/exit. The new area visible on the world map (since the teleport moved east) was added.