Made the old hut exterior look more like ruins.
To make the old hut look better, and more in-line with the interior, I changed it so it looks more like ruins than a copy-pasted building.
Added a chest that will make the player interested in entering the ruins. Once they've entered the interior map (M90) and opened the chest, the event will update to an opened chest (i.e. showing that you've already been here).