Fixed issue with west door being locked.
The door to the west will now be open if the player enters after first
killing the goblin, and then gets the quest to head west.

The issue was with the self switch B for the door not being removed unless
the relevant quest had been picked up before the goblin got killed.

B will now always be turned Off for the door to the west, after the goblin
was killed.

Added a wait for movement completion for door animation if it opens, to
avoid ugly animation.