Lowered damage to clothes during rape attack.
Checks are now if...

  player.agi > enemy.str + d+enemy.str

Damage done is now calculated by the following...

  base_dmg = (enemy.str * d+enemy.str) + d+enemy.atk

i.e. base_dmg can be up to str^2 + atk
base_dmg is not allowed to exceeds 1/5 of the clothes hp. This is OK for normal goblins and the cotton robe, but I'm not sure how well it is when we go further down the line with armors like mythril (80k vs 1.1M hp).

*d+ is a dice roll, min 0, top is the value right of the plus.