Fixed regression bug with rape skill and dog Garnet.
When the rape skill is called, the target is checked with is_nude? rather
than switch 90 (nude).

Dog Garnet doesn't trigger the nude switch when no equipment is worn.
Breaking the script as the game tries to retrieve the max hp of an armor
that doesn't exist.

I don't remember why dog garnet doesn't trigger switch 90. It was probably to solve page condition issues. We should evaluate if this should remain the default behaviour.