Made sp_cost return an int, rather than a float.
sp_cost contains the number of time a spell can be cast, given the current mana pool. This makes sure that this value is an integer i.e. you can cast this spell 2 times, rather than 2.4321456731345 times.

The reason this was changed is mostly for aesthetics, but also for a test to see if it has any impact on performance. Testing using kellessdee's Frame Rate Display, suggests that this change didn't have a noticeable impact displaying it as a float or an int.