BlizzABS hud will now check for SP int changes.
Updated the hud system to look for changes in whole numbers, rather than decimals, when deciding to redraw the bars. This will make the script call the draw function, hopefully improving performance.
Also removed the force update of the sp when hp is changed, not sure if it's needed, but I didn't see anything that suggested that.