Started updating bramble woods east.
Updated the style to have walls to the north and south. So far the
teleports have mostly remained intact, and you don't have to update
surrounding maps.

Moved a goblin event down (that shows up after clearing the caves), to
avoid getting him stuck.

Issues with some of the south tiles that seems to lack priority and
passage data.