Added methods to game_party for getting aggregate data.
You can now get some stats on the inventory, mostly related to money. Check liquidity and what is locked in usage (equipped).
Intended to be used when checking how wealthy a player or her party is, and maybe if there's someone willing to barter for some liquidity in special circumstances.

One possible use is to have a merchant offer money for an item that's equipped (i.e. in use by the player), when she doesn't have things to sell in her inventory, or money at hand to buy things. Another possibility could be to implement it when playing a game, so the equipment can be used to cover the bet.