Updated wolf clearing map.
Changed the aesthetics of the map where you get the status, to a more denser forest. Overdid some of decoration down the path, and you can actually walk through the dense forest; but I think it looks a little bit better now.
Updated the teleport from Remus to place the player on the new path.