Fixed BGM on world map.
Changed the settings of the map to play 018-Field01, rather than having a parallel event trigger it. This was done by right clicking on the map in the editor's map list, and selecting the music under the Auto-Change BGM option.

The reason you don't want to use a parallel or autorun event, is that the event by default will try to run the event again once it done.
I.e. for every frame you're on the game map, the game will try to re-play 018-Field01, which isn't needed and has a negative impacts on performance.

If you wish to use events to change music through an autorun or parallel event, make sure that the event page doesn't repeat by using self switches, or use the erase event command. This is true for all usage of autorun or parallel events, and failure to adhere to the principle of breaking the loop may result in unexpected behavior in the game, and poor performance.

For music and background sound effects, modify the properties of the map, since it's specifically made to manage this functionality.