Fixes some issues with waitress guild quest.
 * Fixed waitress dialog starting before speaking to the bartender.
 * Fixed issue with Garnet not remembering who she has spoken to.
 * Added stubs for getting a maid uniform.
 * Made the old man event require the maid uniform to trigger.

This is a fix to solve some of the old issues related to the waitress guild quest in Dali. The main issue is that the system implemented history/remember what you do, rather it takes any action as the progression of the quest. Something that makes it possible to spam one action over and over again to finish the quest (e.g. you could speak to one person 8 times to clear the quest).

This fix implements a simple binary list for history (for storing persons
visited and persons reported). This is simpler than using self switches
since the inside of the pub is rather complex for the events (could have
required a new map completely) and doesn't require more variables or
switches to clutter the global scope.

Variable 50 now is now a binary array with different index values for different events in the quest.
Index 1 : Quest started.

Index 2 : Spoken to elderly man
Index 3 : Spoken to distant traveler
Index 4 : Spoken to Captain
Index 5 : Spoken to Solider 1 (the one to the north)
Index 6 : Spoken to Solider 2 (the one to the south)
Index 7 : Spoken to Demeter
Index 8 : Spoken to John

Index 12 : Relayed elderly man's order to bartender
Index 13 : Relayed distant traveler's order to bartender
Index 14 : Relayed Captain's order to bartender
Index 15 : Relayed Solider 1's order to bartender
Index 16 : Relayed Solider 2's order to bartender
Index 17 : Relayed Demeter's order to bartender
Index 18 : Relayed John's order to bartender

The pimp and his friend is not in the list, since Garnet will refuse to speak to them. This is the old behaviour and the quickest to implement for now. Index 9 and 10, 19 and 20; has been reserved just in case we want to have them more active during the quest.

To check any of these values you can create a conditional with the following script:

    $game_variables[50] & ( 1 << index ) != 0 # See if index is set

$game_variables[50] being variable 50 (Work) and index being the number to check, e.g. if you wish to check if the captain has been spoken to you type the following

    $game_variables[50] & ( 1 << 4 ) != 0 # See if the captain has been spoken to

To set the variable you can use the script event command, and the |= operator, e.g.

    $game_variables[50] |= ( 1 << 7 ) # Set Demeter as spoken to

The quest is considered complete when index 12-18 is turned on when speaking to the bartender.