Adds maid uniform.
* Defines main armor in armor database.
* Sets Garnet's sprite when maid uniform is worn.

The definition is a quick stub, not sure what kind of values it should have. Probably equal to the cotton robe, since it's supposed to be easy to get through the waitress quest. Currently, it only has the value of 1000 set, and nothing else.

There's a persona armor Zony made in the Assets folder, but the size is currently wrong, Garnet needs to be removed, and the background needs to be replaced (needs to be transparent).