Makes maid uniform buyable.
* Added a block to avoid NPCs walking behind the counter and blocking access to the bartender.
* Upped the base price of the maid uniform to 150.

You can now buy the uniform after the waitress quest. Also scripted sequence if the player sells the armor before finishing the quest.

If the player runs of and sells the uniform, and can't repay the cost (bound to the cost of the armor in the armor definition), the debt will be added to the pimp's RPS debt. It's easier to leverage what we already have when it comes to the debt system, rather than implementing something completely new for the bartender. Alternatively we can have some sort of debt specifically for the guild, who would reimburse the quest givers for Garnet's failures.

Pricing was set to be dynamic, so if the price of the maid uniform is altered in the armor definition (e.g. to something considerably higher), the event should still work as intended. Just remember that the player can leave the quest at any time, so don't make it too good, or the player will just start the waitress quest, and then just continue on with the rest of the game without finishing it.

The event implements a simple script that makes the player randomly select an article of clothing from the inventory, if one exists (used when returning the uniform). I think this has been done before, but I couldn't find it, so it may be a good idea to move this script to a common event for easy access.