Changed switch for the Potts' end.
Moved to using switch 86, since 10 is for the waitress quest. If the player at any times entered the Potts building while doing the waitress quest, they would trigger the bad end, event if they weren't supposed to.

Switch 86 named to "Potts' slave", and changes were made to the bad end trigger (Potts in map 1) and in the Potts building (map 4).