Makes the player return any pig food they have.
When the player leaves the pig pen in Erica's Farm, the game will now check if she is carrying any pig food with her. If she does, there will be a short sequence where she returns the smelly food.

This is the same behaviour that exists in the Dali farm.