Removes scripts binary file.
The binary file was previously used to force users to run the batch
files, even if they didn't read the readme (which many didn't). But this
was problematic, since it constantly changes during development. The
issue is that a developer may accidentally, or by the usage of his tool,
commit the file anyway.

If the changes are committed, it breaks the function of the file, and
has to be reverted.

Since some Git tools (like Github for Windows) by default want you to
commit all the changes you've made; it's only a matter of time until
this will cause a major issue.

This commit aims to "fix" this, by removing the file completely from
being tracked. Further changes needs to be made so we can catch those
not running the batch files in some way, but this was clearly not the
correct way of doing things.