Adds a bladder to the player.
The player will now have a bladder that can be filled as the game is played.

Turn ON Switch 173 to make the feature available.
The current bladder value (how filled it is, 0-100) can be found in variable 21
When the bladder becomes full, it will trigger switch 174 (which can be used to trigger common events or events on the map)

The current update rate is 720 frames, with 1 being added to the 0-100 range. After 72000 frames the bladder will be filled. Since the game plays at 40fps, this means that you will have a filled bladder after 30 minutes (72000frames/40fps/60spm). Modify the FRAME_UPDATE constant to change the number of frames, or MODIFIER to alter how much is changed at a time (2.0 = double, 0.5 = half).

If you change the MODIFIER you may wish to set FORCE_INTEGER = FALSE, so the value isn't rounded off into no change at all. You may want to keep FORCE_INTEGER on, if you intend to show the value in any way on the screen (besides the bar).