Updates Keni Family home interior.
Redid the interior to be more in line with the rest of the maps.
Adds 7*2 tile vertical buffer and 9*2 horizontal tile buffer, to make the movement more free within the map.
Updated teleport from Dali North.
2nd floor remains closed (where I'm guessing the adults sleep).
Most of the interior content is reused from the old map, to keep the style. Weapons were removed, since I'm not sure if they served a purpose or not.