Updated condition change -> event updates.
Added wrappers for the event handlers, so you may now call it with anything that can use a call method. You still need a wrapper of some sort for the marshalling, but it doesn't have to be anything advanced.

Map events will now also listen to variable changes, besides switches which existed before.