Makes bladder switch event driven.
Instead of updating the switch with every fixed bladder change tick, the switch will listen to any changes to the bladder, and modify its own value appropriately. Note that his may have some unintended consequences if it fires updates outside of the current interpreter iteration (i.e. things may disappear, or event may exit half way through), but I haven't seen any of that yet.

This makes it possible to script changes to the bladder variable without having to wait up to 18 seconds for the game to realize this.