Reorders refresh update.
Moves the on-demand refresh to the end of the map update. This was the only way to fix the bug with the knockout manager not firing, and I have no idea why.
Adds extra checks to make sure the events doesn't get refreshed a second time if there was a forced refresh of all events on the map.