Fixes dialog jump when starting book quest.
Removed the proxy, since it made the NPC trigger two times when talked
to from the north (facing south). The issue is that the player stands
on one trigger (the proxy) activating it, and at the same time
activating the event in-front of them. The result being that the event
is triggered two times> directly, and indirectly.

The reasons the proxy was there in the first place, was likely due to
priority and clipping (passage) issues, where the player wasn't able to
get close enough to the NPC to trigger it (see various merchants behind
desks). Changes to the tileset has since made this obsolete, and at the
same time introduced the double trigger bug.

The book quest will now start as the player talks to the NPC the first
time, rather than having to trigger the dialog a second time. If the
player talks to the NPC without any books, they will trigger the second
dialog where Garnet tries to convince the NPC to lend her books without
returning them.

If the player tries to trigger the NPC a third time without having any
books, they will be reminded of the quest through a short dialog.