Fixes training at dojo.
Removes the ability to punch the pole, dummy, and stone pillar; if the
player's HP is less than the damage done by the action.

For some reason the knockout state doesn't kick in, suggestion there's
something wrong with the damage dealing ability of the event commands.

Damage done by enemies still work as intended, and will trigger the
knockout state.