Fixes issue with knockout CE not triggering.
If the is a knockout manager on the map that doesn't trigger as the
player enters their kockout state (if the conditions aren't met by the
page), the game will now use the default knockout manager to handle the
event.

The change to the code is a check if the list of commands retrieved is
empty instance, by the first command having the code 0. Technically the
command with code 0 does nothing in the engine, and is ignored; but
since it can only be created when the command object is constructed
(default value) as a new command is created (commands with code 0 can't
be made in the editor), we assume that having a command with code 0 as
the first command, means that the list is an empty construct.

Another solution would have been to check the length of the list of
commands. Since there's always and empty (code 0) commands at the end
of the list, we know that any list with just 1 command is the same as
having no commands. This is not used since we can't be sure that the
commands passed actually follow this style of having an empty command
at the end (since it isn't required by the engine), which would limit
the usability of the function.