Dedicates 22 maps for Clairia and Hikari.
Hikair and Clairia is given 10+1 empty maps each, do avoid conflicts with map indices.

When you need a new map, you drag one of the available maps to the location where you want it.

When you start reaching the end of your buffer (make sure you always have a buffer) announce that you'll be adding more maps to your bucket, and send a pull request. It's first come, first serve; so if you're trying to modify a map already taken it will (in the future) be moved.