Adds taberenc's pig image to pig animation.
Garnet's pig animation will now use the naked one by default, and only
use the original images when Garnet is wearing a cotton robe.

This is not the best way of doing it, since it will do lookups with
every frame, which we don't need. Unfortunately, since we use the same
image for animation, we need to solve it like this, or alter the way we
do animations alltogether.