Updates on spelling and improving current events and add new characters
1st I've gone over a lot of the talking and found I have a habit of put
Garent and not Garnet this has been fixed.

I've added two new characters to Dali Village 001 at the table on the
west side Lance and Alice Noels father and sister and they will only
appear if switch 0444 Tranced is on say yes to them will turn Tranced
off and a new switch Lance told 0447 will be turned on this is to make
another line of events to happen which will be added later.

Next i have improved my event on bramble woods 033 so that it is no
longer on just one spot this needs switch 0039 on to play and it will
turn switch Seen Noel on 0445.

For the old hut 239 for event 006 I've improved it at parts for who
Garnet is Tranced and added a self switch A to turn the event off rather
than just using switch 0444 Tranced also for this event I found a bug
that after the event for turning Trance on the player would be unable to
move if they were shy this has been fixed for now by keeping the players
colthes on I'll be working on a better fix soon.

Added a 1st floor to the hut 241 this can only be acessed if switch 0444
Tranced is on but has no events right now.