Improvment on event and fix for a bug
Just more updates for old hut 239 making the event cleaner as when I had
Garnet walking out of the hut she would have her armour on and then take
it off again this is now fixed.

The 2nd problem is if the naked ok switch was not on when Garnet got to
the world map she would want to put her armour back on but the menu
would not load so you couldn't move right now to fix that the event now
gives 200 to variable 0049 pervert for now other than making the events
require it i'm not sure on how else to fix it right now.