New Event Added for Tranced Switch
Add a new switch called plant clone 0448 which is turned on by this new
event the main idea behind this is for a few mini events for later.

The changes to old hut map 239 is a autorunner for the event added this
will work if Garnet enters the hut with the Tranced switch on 0444.

As for the event this will tigger 1 of 2 scences depending how high the
pervet variable 0049 is 350 or greater i'm not to sure on if that might
be too high or low yet. I've used a self switch so the event does not
happen again.

Once either scene this finshed this will lead to the basement map 240
which is another autorun event for the tranced switch 0444 also it can
only be accessed if tranced is on.

The basement event can be a bit buggy with the animeation for little
parts and I can't seem to find the reason why as sometimes it works fine
then another it messes up a little. I'll look into this again later as
the effect is small for now. This is where the new switch plant clone
0448 is turned on.

One of my main thoughts about my events as a whole is I'm not sure if
I'm making them too long this event is the longest of what I have done
so far some feed back on this point would be nice thanks.

Lastly is the 1st floor map 241 this is based on the new switch plant
clone 0448 and you can only go up there for now if its on.

This has a example of the mini events I mean for the plant clone switch
0448.