New event for Forced Tranced
New event for switch 0446 Forced Tranced on map 239 Old Hut

So new event for Garnet walking into the hut at the wrong time.

Main problem is with Leon for this update I know it doesn't look that
great but for that graphic for Leon events 007-fighter07 is the closest
looking one that is wearing something so thats why it is used if someone
that is good with that kinda stuff could help with that it be great if
not I'll look into learning bout that stuff later.

I also now made it so those with forced tranced on can now enter the
basement.