Forced Tranced Event for the basement
Forced Tranced switch 0446 OW_Basement_2 map 240

This is the same Event from the old hut map 239 just carrying.

Its finshed as far as the graphics and talking I just need to go back
and add sounds to the event.

I also added Leon to the basement now for those with Forced Tranced on.