Just a quick fix to a few things I noticed
For the Forced tranced event 0446 in the basement I had the text show at
the top of the screen so the event wasn't blocked by text but I forgot
to change it back to normal at the end and so it would affect other text
boxs later in the game this has been fixed now.

And the way to the basement for Tranced Force was left on the player to
use the action button rather than just walk over it this has now been
fixed as well.