Moved static engine files to a seperate folder.
Since many Git clients have a tendency to recommend committing that
all files modified, some unfortunate consequences follow.

In this case, the game is based on an older version of the RPG Maker
Engine (1.02) which has a seperate dll (which ships with the game),
and while newer versions of the engine is backwards compatible with
1.02, the opposite is not true.

The issue has been that developers have committed their changes with
and updated definitions of what dll files to use, and what editor
is compatible. This has basically broken it for players and other
developers since unless they have the latest editor or RTP installed
will get an error message they can't easily resolve.

This solution is to keep the relevant files separate from where the
editor and game reads and modifies them.

The files are copied in as the user runs game.bat or editor.bat.
Even if the files are modified on the developers system, the
changes will not be sent upstream by default.

If there's a need to update the static files, just copy the relevant
files from the Game folder into Src\Required. All files in the
Src\Required folder will be used to overwrite whatever is in the Game
folder, and this can be used for other fixes as well.