1st stage of map updates
As far as I know, all maps that needed updating for the new tilesets are
done. I haven't done a complete check of the maps so there are probably
minor issues with some of the maps. Expect at least two more updates as
there are still several maps that I'd like to combine, and the tileset
database needs corrections made for game play. To keep everthing clean,
I removed the old tileset entries and their related files.