Made the access in the Dali crypt more clear.
- Added animation for the automatic opening of the first door.
- Added animation for the opening of the boss door.
- Changed the look of the first door to be more "tomb" like.
- Changed the look of the boss door to be more "normal door" like.

Since there's some confusion to what door the key goes to, I changed
the first door to look more like a "tomb door" and changed the text so
it isn't implied that you should "unlock" it (which would be the
obvious use of the key you get right next to it). A short paragraph was
also added for players that has the key in their inventory, so it's
clear that the key isn't the solution to this "puzzle".

The boss door was changed to a normal wooden door to make it clear that
it has a different solution for unlocking.

Changed the animation of the door unlocking, so the entire thing fades
out after it has been opened. This is so we don't have the ugly black
boxes that comes with the open doors in normal cases. A better solution
could be to create doors with alpha-transparency, so they'd merge
better with their surroundings.