Named events and updated remote switch calls.
Updated the events so they should work better with the new merged map.
Since the the old maps were separate the calls to toggle remote
switches were firmly rooted in the index position of the event (and to
some extend the index of the map) they needed rewriting to be rewritten
for a new event indices (locations).

Rather than having a fixed index reference, events were named and are
now called based on their event names.

Unfortunately we can't easily name what the switches (A/B/C/etc) means
for the event, but at least it's clear what event we call when we
modify them.