Updated passage to blocking for doll in Garnet Tile 55.
Fixes #136

If the tile used for en event can be threaded over willy-nilly, the event
won't trigger unless you're right on-top of it (standing on it). That's an
issue if the event you want to trigger is on something that that the
player can't traverse, like the base of a pillar, which was the issue with
the quest for a certain doll.